requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-cors==4.0.0
python-docx==0.8.11
docx2pdf==0.1.8
reportlab==4.0.4
email-validator==2.1.1
bcrypt==4.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
PyJWT==2.8.0
marshmallow==3.20.1
sqlalchemy==2.0.25
docxtpl==0.16.4
